+++ core/lib/multivm/src/versions/vm_m6/oracle_tools.rs
@@ -13,7 +13,7 @@ use zk_evm_1_3_1::witness_trace::DummyTracer;
 
 /// zkEVM requires a bunch of objects implementing given traits to work.
 /// For example: Storage, Memory, PrecompilerProcessor etc
-/// (you can find all these traites in zk_evm crate -> src/abstractions/mod.rs)
+/// (you can find all these traits in zk_evm crate -> src/abstractions/mod.rs)
 /// For each of these traits, we have a local implementation (for example StorageOracle)
 /// that also support additional features (like rollbacks & history).
 /// The OracleTools struct, holds all these things together in one place.