+++ core/lib/zksync_core/src/api_server/web3/namespaces/mod.rs
@@ -4,17 +4,11 @@
 mod debug;
 mod en;
 pub(crate) mod eth;
-mod eth_subscribe;
 mod net;
 mod web3;
 mod zks;
 
 pub use self::{
-    debug::DebugNamespace,
-    en::EnNamespace,
-    eth::EthNamespace,
-    eth_subscribe::{EthSubscribe, SubscriptionMap},
-    net::NetNamespace,
-    web3::Web3Namespace,
-    zks::ZksNamespace,
+    debug::DebugNamespace, en::EnNamespace, eth::EthNamespace, net::NetNamespace,
+    web3::Web3Namespace, zks::ZksNamespace,
 };